+++ lib/core/util.js
@@ -6,7 +6,6 @@ const { IncomingMessage } = require('node:http')
 const stream = require('node:stream')
 const net = require('node:net')
 const { Blob } = require('node:buffer')
-const { markAsUncloneable } = require('node:worker_threads')
 const nodeUtil = require('node:util')
 const { stringify } = require('node:querystring')
 const { EventEmitter: EE } = require('node:events')
@@ -817,12 +816,6 @@ function errorRequest (client, request, err) {
   }
 }
 
-function maybeMarkAsUncloneable (target) {
-  if (markAsUncloneable !== undefined) {
-    markAsUncloneable(target)
-  }
-}
-
 const kEnumerableProperty = Object.create(null)
 kEnumerableProperty.enumerable = true
 
@@ -882,7 +875,6 @@ module.exports = {
   isFormDataLike,
   serializePathWithQuery,
   addAbortListener,
-  maybeMarkAsUncloneable,
   isValidHTTPToken,
   isValidHeaderValue,
   isTokenCharCode,
