+++ lib/web/eventsource/eventsource.js
@@ -3,13 +3,13 @@
 const { pipeline } = require('node:stream')
 const { fetching } = require('../fetch')
 const { makeRequest } = require('../fetch/request')
-const { webidl } = require('../fetch/webidl')
+const { webidl, maybeMarkAsUncloneable } = require('../fetch/webidl')
 const { EventSourceStream } = require('./eventsource-stream')
 const { parseMIMEType } = require('../fetch/data-url')
 const { createFastMessageEvent } = require('../websocket/events')
 const { isNetworkError } = require('../fetch/response')
 const { delay } = require('./util')
-const { kEnumerableProperty, maybeMarkAsUncloneable } = require('../../core/util')
+const { kEnumerableProperty } = require('../../core/util')
 const { environmentSettingsObject } = require('../fetch/util')
 
 let experimentalWarned = false
