+++ lib/web/fetch/formdata.js
@@ -1,8 +1,8 @@
 'use strict'
 
 const { iteratorMixin } = require('./util')
-const { kEnumerableProperty, maybeMarkAsUncloneable } = require('../../core/util')
-const { webidl } = require('./webidl')
+const { kEnumerableProperty } = require('../../core/util')
+const { webidl, maybeMarkAsUncloneable } = require('./webidl')
 const { File: NativeFile } = require('node:buffer')
 const nodeUtil = require('node:util')
