+++ lib/web/fetch/headers.js
@@ -3,13 +3,13 @@
 'use strict'
 
 const { kConstruct } = require('../../core/symbols')
-const { kEnumerableProperty, maybeMarkAsUncloneable } = require('../../core/util')
+const { kEnumerableProperty } = require('../../core/util')
 const {
   iteratorMixin,
   isValidHeaderName,
   isValidHeaderValue
 } = require('./util')
-const { webidl } = require('./webidl')
+const { webidl, maybeMarkAsUncloneable } = require('./webidl')
 const assert = require('node:assert')
 const util = require('node:util')
