+++ lib/web/fetch/request.js
@@ -22,8 +22,8 @@ const {
   requestCache,
   requestDuplex
 } = require('./constants')
-const { kEnumerableProperty, maybeMarkAsUncloneable, normalizedMethodRecordsBase, normalizedMethodRecords } = util
-const { webidl } = require('./webidl')
+const { kEnumerableProperty, normalizedMethodRecordsBase, normalizedMethodRecords } = util
+const { webidl, maybeMarkAsUncloneable } = require('./webidl')
 const { URLSerializer } = require('./data-url')
 const { kConstruct } = require('../../core/symbols')
 const assert = require('node:assert')
