+++ lib/web/fetch/response.js
@@ -4,7 +4,7 @@ const { Headers, HeadersList, fill, getHeadersGuard, setHeadersGuard, setHeaders
 const { extractBody, cloneBody, mixinBody, hasFinalizationRegistry, streamRegistry, bodyUnusable } = require('./body')
 const util = require('../../core/util')
 const nodeUtil = require('node:util')
-const { kEnumerableProperty, maybeMarkAsUncloneable } = util
+const { kEnumerableProperty } = util
 const {
   isValidReasonPhrase,
   isCancelled,
@@ -18,7 +18,7 @@ const {
   redirectStatusSet,
   nullBodyStatus
 } = require('./constants')
-const { webidl } = require('./webidl')
+const { webidl, maybeMarkAsUncloneable } = require('./webidl')
 const { FormData } = require('./formdata')
 const { URLSerializer } = require('./data-url')
 const { kConstruct } = require('../../core/symbols')
