+++ lib/web/fetch/webidl.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const { types, inspect } = require('node:util')
+const { markAsUncloneable } = require('node:worker_threads')
 const { toUSVString } = require('../../core/util')
 
 const UNDEFINED = 1
@@ -699,6 +700,13 @@ webidl.converters['record<ByteString, ByteString>'] = webidl.recordConverter(
 
 webidl.converters.Blob = webidl.interfaceConverter(Blob)
 
+function maybeMarkAsUncloneable (target) {
+  if (markAsUncloneable !== undefined) {
+    markAsUncloneable(target)
+  }
+}
+
 module.exports = {
-  webidl
+  webidl,
+  maybeMarkAsUncloneable
 }
