+++ lib/web/websocket/events.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const { webidl } = require('../fetch/webidl')
-const { kEnumerableProperty, maybeMarkAsUncloneable } = require('../../core/util')
+const { webidl, maybeMarkAsUncloneable } = require('../fetch/webidl')
+const { kEnumerableProperty } = require('../../core/util')
 const { kConstruct } = require('../../core/symbols')
 const { MessagePort } = require('node:worker_threads')
